# openapi.yaml
openapi: 3.1.0
info:
  title: IREC — API v1
  version: 1.0.0-draft.1
  description: |
    **BOZZA** — contratto di integrazione Mind → IREC, da concordare con il team Mind.
    Questo file è la fonte di verità dell'integrazione (contract-first).

    Convenzioni:
    - Autenticazione: call-token coniato da Mind (JWT firmato, verifica JWKS lato IREC).
      Claim richiesti: `sub`, `tenant_id`, `entitlement`, `scope`, `aud: "irec"`, `jti`, `exp`.
    - `x-correlation-id`: ricevuto da Mind e ri-emesso nei log di IREC.
    - Mutazioni idempotenti via header `Idempotency-Key`.
    - Run lunghe: risposta `202 { run_id }`, poi poll su GET.
    - Versioning `/v1`: solo cambi additivi; policy di deprecation da definire.
    - Gli endpoint a supporto dei tool dell'agente (addendum agentico §6:
      posizioni, storico solleciti, pausa/riprendi, ecc.) verranno aggiunti
      in modo additivo in revisioni successive di questa bozza.
servers:
  - url: /v1
security:
  - callToken: []
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      summary: Elenco fatture del tenant, filtrabile per stato
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: status
          in: query
          description: >-
            Stato operativo della fattura. `scaduta` non è uno stato
            persistito ma un filtro derivato (data_scadenza passata e
            fattura non ancora saldata).
          schema:
            type: string
            enum: [gestione, pausa, saldata, insoluto, scaduta]
        - name: as_of
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Fatture del tenant
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
        '403': { $ref: '#/components/responses/EntitlementError' }
  /aging:
    get:
      operationId: getAging
      summary: Bucket di aging del credito (a_scadere, 0-30, 31-60, 61-90, 90+)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: as_of
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Aging del portafoglio del tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgingReport'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /usage:
    get:
      operationId: getUsage
      summary: Consumo per il billing (aggregato da Mind)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Metriche di consumo del tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usage'
  /portfolio:
    get:
      operationId: getPortfolio
      summary: KPI del portafoglio (affidato = recuperato + da recuperare + a recupero)
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: KPI del tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Kpi' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/EntitlementError' }
  /positions/{position_id}:
    get:
      operationId: getPosition
      summary: Dettaglio di una posizione con le sue fatture
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: position_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Posizione
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /invoices/{invoice_id}/history:
    get:
      operationId: getInvoiceHistory
      summary: Cronologia delle comunicazioni di una fattura
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: invoice_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Storico comunicazioni
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Storico' }
        '404': { $ref: '#/components/responses/NotFound' }
  /invoices/{invoice_id}/next:
    get:
      operationId: getNextCommunications
      summary: Step programmati futuri di una fattura
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: invoice_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Prossimi invii
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProssimiInvii' }
        '404': { $ref: '#/components/responses/NotFound' }
  /communications/{communication_id}/explain:
    get:
      operationId: explainCommunication
      summary: Perché una comunicazione è (o non è) partita
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - { name: communication_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Codice di spiegazione
          content:
            application/json:
              schema:
                type: object
                required: [comunicazione_id, codice]
                properties:
                  comunicazione_id: { type: string }
                  codice: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    callToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Call-token coniato da Mind, verificato da IREC via JWKS.
  parameters:
    CorrelationId:
      name: x-correlation-id
      in: header
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: >-
        Call-token assente o non valido. `code`: missing_token, invalid_token,
        token_expired, invalid_audience, unknown_key, missing_tenant.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    EntitlementError:
      description: >-
        Autorizzazione insufficiente. `code`: entitlement_missing (nessun
        entitlement IREC), scope_missing (scope richiesto assente dal token).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Risorsa non trovata
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
    Kpi:
      type: object
      required: [affidato, recuperato, da_recuperare, passato_a_recupero]
      properties:
        affidato: { type: string, format: decimal }
        recuperato: { type: string, format: decimal }
        da_recuperare: { type: string, format: decimal }
        passato_a_recupero: { type: string, format: decimal }
        fatture_per_stato: { type: object, additionalProperties: { type: integer } }
        posizioni_aperte: { type: integer }
        posizioni_chiuse: { type: integer }
    Position:
      type: object
      properties:
        id: { type: string }
        cliente: { type: string }
        piva_cf: { type: string }
        stato: { type: string }
        importo_totale_residuo: { type: string, format: decimal }
        fatture:
          type: array
          items: { $ref: '#/components/schemas/Invoice' }
    Comunicazione:
      type: object
      properties:
        id: { type: string }
        canale: { type: string }
        template: { type: string }
        stato:
          type: string
          enum: [programmata, inviata, annullata, fallita, saltata]
        programmata_per: { type: string, format: date-time }
        inviata_at: { type: [string, 'null'], format: date-time }
        esito_recapito: { type: [string, 'null'] }
    Storico:
      type: object
      properties:
        fattura_id: { type: string }
        comunicazioni:
          type: array
          items: { $ref: '#/components/schemas/Comunicazione' }
    ProssimiInvii:
      type: object
      properties:
        fattura_id: { type: string }
        prossimi:
          type: array
          items: { $ref: '#/components/schemas/Comunicazione' }
...
    Invoice:
      type: object
      properties:
        id: { type: string }
        numero: { type: string }
        cliente: { type: string }
        piva_cf: { type: string }
        data_emissione: { type: string, format: date }
        data_scadenza: { type: string, format: date }
        # Gli importi viaggiano come stringhe decimali: il dominio è
        # finanziario e un double introdurrebbe errori sui centesimi.
        importo: { type: string, format: decimal, examples: ['1220.00'] }
        importo_residuo: { type: string, format: decimal, examples: ['0.00'] }
        stato:
          type: string
          enum: [gestione, pausa, saldata, insoluto]
    AgingReport:
      type: object
      properties:
        as_of: { type: string, format: date }
        buckets:
          type: array
          items:
            type: object
            properties:
              label: { type: string, examples: ['0-30', '31-60', '61-90', '90+'] }
              importo: { type: number }
              numero_fatture: { type: integer }
    Usage:
      type: object
      properties:
        tenant_id: { type: string }
        period_from: { type: string, format: date }
        period_to: { type: string, format: date }
        metrics:
          type: object
          additionalProperties: { type: number }